pyproject.toml
[tool.poetry]
name = "pupil_check_back"
version = "0.1.0"
description = ""
authors = ["Marcan <marcu-andria.battesti@jellysmack.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.88.0"
dependency-injector = "^4.41.0"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
